move utils.ts from src/assets -> src/util.ts
