+++ electron/ui/src/components/NetworkMap/MapEditor.tsx
@@ -3,7 +3,7 @@ import { Box, Button, TextField, IconButton, MenuItem, Typography, Stack, Toolti
 import { InputAdornment, InputLabel, Select, FormControl } from '@mui/material';
 import { useMapValues } from '../../context/MapContext';
 import type { CoordinateTuple, SelectedNodeState, MapEditorNode } from '../../types';
-import { NetworkNodeTypes, checkIfNameIsUnique, useKeyDown } from '../../assets/utils';
+import { NetworkNodeTypes, checkIfNameIsUnique, useKeyDown } from '../../util';
 import CloseIcon from '@mui/icons-material/Close';
 import AddCircleIcon from '@mui/icons-material/AddCircle';
 import EditIcon from '@mui/icons-material/Edit';
