+++ electron/ui/src/views/ModelResults/ModelResults.tsx
@@ -4,7 +4,7 @@ import { Box, Grid, LinearProgress, Button } from '@mui/material';
 import RefreshIcon from '@mui/icons-material/Refresh';
 import SankeyPlot from './SankeyPlot';
 import KPIDashboard from './KPIDashboard';
-import { TerminationConditions } from '../../assets/utils';
+import { TerminationConditions } from '../../util';
 import NetworkDiagram from '../../components/NetworkDiagram/NetworkDiagram';
 import DataTable from '../../components/DataTable/DataTable';
 import FilterDropdown from '../../components/FilterDropdown/FilterDropdown';
